Fix most oustanding breakage in wasm_backend from 0xc (#4677)
* Cast pointer to int in test_alloca to avoid a warning
* Remove redundant assert in test_autodebug, remove disable-in-v8 code
* Filter non-func imports from s2wasm import parsing, to avoid considering memory imports as undefined symbols
* Explicitly prevent inlining in test_stack_align to make what's being tested clearer

+++ tests/core/test_alloca.c
@@ -9,6 +9,6 @@ int main(int argc, char **argv) {
   assert(((int)pc) % 4 == 0);
   pc2 = (char *)alloca(4+argc);
   assert(((int)pc2) % 4 == 0);
-  printf("z:%d*%d*%d*\n", pc > 0, (int)pc, (int)pc2);
+  printf("z:%d*%d*%d*\n", (int)pc > 0, (int)pc, (int)pc2);
   return 0;
 }
